Separate details view (inside) from opening the whole simulator
