+++ src/app/features/simulator/simulator/simulator.component.html
@@ -38,7 +38,7 @@
     <g text-area-svg [(text)]="conv.title" [singleEdit]="true" [w]="600" [h]="25" [x]="0" [y]="0"></g>
     <g authorG [author]="conv.author"></g>
     <g cp *ngFor="let convP of conv.conversationPartners" [conversationPartner]="convP"></g>
-    <g preknowG *ngFor="let pre of conv.author.preknowledge" [info]="pre" [showTrust]="true" (chooseElem)="simulationDialogueService.openInfoTrusts(pre, conversation, trustFallbacks)"></g>
-    <g msgG *ngFor="let m of conv.author.messages" [msg]="m" [msgTiming]="m.timing" [showInfoTrusts]="true" (chooseInfo)="simulationDialogueService.openInfoTrusts($event, conversation, trustFallbacks)"></g>
+    <g preknowG *ngFor="let pre of conv.author.preknowledge" [info]="pre" [showTrust]="true" (chooseElem)="infoTrustDetailsService.openInfoTrusts(pre, conversation, trustFallbacks)"></g>
+    <g msgG *ngFor="let m of conv.author.messages" [msg]="m" [msgTiming]="m.timing" [showInfoTrusts]="true" (chooseInfo)="infoTrustDetailsService.openInfoTrusts($event, conversation, trustFallbacks)"></g>
   </svg:g>
 </ng-template>
